added ping to server and cloud functions

+++ domain/schema_registry.go
@@ -2,8 +2,8 @@ package domain
 
 type SchemaRegistry struct {
 	Host 			string
-	Port 			int64
-	NumOfSchemas	int64
+	Port 			int
+	NumOfSchemas	int
 	Schemas 		[]Schema
 }
 
@@ -11,5 +11,5 @@ type Schema struct {
 	ID 			float64
 	Name 		string
 	NameSpace 	string
-	MaxVersion 	int64
+	MaxVersion 	int
 }
